+++ glue/core/tests/test_data_collection.py
@@ -12,7 +12,7 @@ from ..component_link import ComponentLink
 from ..data import Data, Component, ComponentID, DerivedComponent
 from ..data_collection import DataCollection
 from ..hub import HubListener
-from ..message import (Message, DataCollectionAddMessage,
+from ..message import (Message, DataCollectionAddMessage, DataRemoveComponentMessage,
                        DataCollectionDeleteMessage,
                        ComponentsChangedMessage)
 
@@ -142,7 +142,8 @@ class TestDataCollection(object):
         assert link in dc._link_manager
 
     def test_catch_data_add_component_message(self):
-        """DerviedAttributes added to a dataset in a collection
+        """
+        DerviedAttributes added to a dataset in a collection
         should generate messages that the collection catches.
         """
         d = Data()
@@ -366,3 +367,22 @@ class TestDataCollection(object):
         dc.merge(x, y)
         assert dc[0].visible_components[0] is x.id['x']
         assert dc[0].visible_components[1] is y.id['y']
+
+    def test_remove_component_message(self):
+
+        # Regression test to make sure that removing a component emits the
+        # appropriate messages.
+
+        data = Data(x=[1, 2, 3], y=[4, 5, 6])
+        self.dc.append(data)
+
+        remove_id = data.id['y']
+
+        data.remove_component(remove_id)
+
+        msg = self.log.messages[-2]
+        assert isinstance(msg, DataRemoveComponentMessage)
+        assert msg.component_id is remove_id
+
+        msg = self.log.messages[-1]
+        assert isinstance(msg, ComponentsChangedMessage)